requirements.txt
# Core framework
tensorflow==2.6.2
keras==2.6.0
tensorflow-addons==0.14.0

# Segmentation models and backbones
segmentation-models==1.0.1
efficientnet==1.0.0
image-classifiers==1.0.0

# Scientific computing
numpy==1.19.5
scipy==1.5.4
scikit-learn==0.24.2
scikit-image==0.17.2

# Image processing
opencv-python==4.5.5.64
Pillow==8.4.0

# Visualization
matplotlib==3.3.4

# Utilities
tqdm==4.62.3
h5py==3.1.0
# Config parsing
PyYAML==6.0.2
